feat(karma): karma config managed automatically
Prior to this commit, any vendor code changes would require manually
adding them to the karma config file. Now, these changes are managed
automatically based on the settings in the Gruntfile.

The `karma/karma-unit.tpl.js` file is a Grunt template that the build
process uses to generate a config file for use during testing. As a
result, the `karma/karma-unit.js` file has been removed.

Unfortunately, there is no easy way to dynamically reload the karma
config file during a watch, so when new vendor files are added to the
build variables in the Gruntfile, we need to restart the watch.

Closes #47.